+++ packages/core/src/index.ts
@@ -8,7 +8,7 @@ import {
 } from './universal';
 import { queryHooks as watchQueryHooks } from './watch';
 import { queryHooks as listenQueryHooks } from './listen';
-export * as Auth from './providers/auth0';
+export * as Auth from './providers/auth';
 export * as Query from './providers/query';
 export { commonHelpers, requestHelpers };
