+++ packages/core/src/universal/hooks/useFeatureFlagSubscription/index.ts
@@ -1,8 +1,8 @@
 import { useMemo } from 'react';
-import { useAuthContext } from '../../../providers/auth0';
 import { useSubscription } from '../useSubscription';
 import type { FeatureFlagsData, FeatureFlagsResponse } from './types';
 import { checkFeatureFlag } from './helpers';
+import { useAuthContext } from '../../../providers/auth';
 
 const FEATURE_FLAGS_SUBSCRIPTION = `
   subscription FeatureFlags {
